tests/wpt/mozilla/tests/mozilla/microdata/tests.html

<!doctype html>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta charset="utf-8">
<html>

<title id="tl" itemtype="title_name"> HELLO</title>
<h1></h1>
<h2></h2>
<p></p>
<input></input>
<form></form>
<li></li>
<ul></ul>
<meta></meta>
<table></table>
<form></form>

<body id="mainBody" itemtype="http://www.google.com">
 <div id="reg" itemtype="hello world">Hi</div>
 <div id="dup" itemtype="hello world world">Hi</div>
 <div id="single" itemtype="hello">Hi</div>
 <div id="extraSpace" itemtype="hello   world">Hi</div>
 <div id="caseSensitive" itemtype="Hello hello" </div>
  <div id="dupNum" itemtype="34562 34562"></div>
  <div id="blank" itemtype=""></div>
  <a id="unicode" itemtype=""></a>
  <button onclick="test()">TEST</button>
</body>

</html>
<script>
 test(function() {
  assert_equals(document.getElementById("single").propertyNames(), "hello", "The method should return 'hello world' ");
 }, "Simple test");
</script>
